Update Transport  Helicopter
The Transport Helicopter script is no persistant.

You can only call the Helicopters that are available.

You can choose if Helicopters can be damaged or be on civilian side.

You can give a cooldown time for after the transport returned to base.

You can decide if destroyed helicopters can respawn or if they are lost for the rest of the mission.