Supply System Rework
Moved everything to functions that is not needed in the loadout folder.

Loadout folder is now reduced to only two files.

CFGFactionEquipment and the createSupplyCrates.sqf

What supply crates are available is moved into CFGFactionEquipment.

Also, CFGFactionEquipment has a new Config Value "factionPath" this removes the old missionConfigValue Faction in the config.hpp.